requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.3
tkcalendar>=1.6.1
openpyxl>=3.1.2
PyPDF2>=3.0.0
matplotlib>=3.5.0
pdfplumber>=0.9.0
cryptography>=39.0.0
aiohttp>=3.8.0
python-dateutil>=2.8.2
PyYAML>=6.0.0
pyinstaller>=5.13.0
schedule>=1.2.0
